Created command groups to set lift
For ease of use, commands were created for each way of setting the lift
on the stacker. Now, rather than calling the SetLift commands directly,
it is better to call the command groups. This helps for ease of writing
and readability. TODO: Measure heights for loading bin and scoring both
on the platform and co-op, then make commands for those heights.